refactor(WIP): Reestrutura o projeto, organizando melhor em pastas e arquivos separados
