+++ simulador/personagens/__init__.py
@@ -1,0 +1,3 @@
+from .personagem import Personagem
+from .guerreiro import Guerreiro
+from .mago import Mago
