+++ simulador/utils.py
@@ -1,4 +1,4 @@
-from models import Personagem
+from .personagens import Personagem
 from os import system, name
 from typing import Dict, Any, Union
 
@@ -17,14 +17,14 @@ def narrador_combate(relatorio: Dict[str, Any], atacante: "Personagem", defensor
     dano_causado: int = relatorio["dano_causado"]
     critico: Union[bool, None] = relatorio.get("critico")
     especial: Union[bool, None] = relatorio.get("especial")
-    motivo_erro_especial: Union[str, None] = relatorio.get("motivo_erro_especial") if atacante.mana_maxima > 0 else None
+    motivo_erro_especial: Union[str, None] = relatorio.get("motivo_erro_especial") if atacante.possui_mana() else None
     especial_cooldown: Union[int, None] = relatorio.get("especial_cooldown")
     
     print(f"======= {nome_duelista1} x {nome_duelista2} =======")
     print(f"Turno: {turno}/100\n")
     
     for duelista in duelistas:
-        if duelista.mana_maxima > 0 and motivo_erro_especial:
+        if duelista.possui_mana() and motivo_erro_especial:
             print(motivo_erro_especial)
             if motivo_erro_especial == "Especial em tempo de recarga" and especial_cooldown is not None:
                 print(f"Faltam {especial_cooldown + 1} turno(os) para {nome_atacante} conseguir utilizar seu especial")
@@ -32,7 +32,7 @@ def narrador_combate(relatorio: Dict[str, Any], atacante: "Personagem", defensor
     
     print(f"{nome_atacante} ataca {nome_defensor}.")
     
-    if motivo_erro_especial or atacante.mana_maxima == 0:
+    if motivo_erro_especial or not atacante.possui_mana():
         print(f"{nome_atacante} utiliza um ataque básico.")
 
     if especial:
